agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
requests==2.32.3
openai==1.51.0
resend==2.3.0
africastalking==1.2.8
hubspot-api-client==9.0.0
playwright==1.47.0
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
langfuse==2.55.0
scipy==1.14.1
numpy==2.1.2
tenacity==9.0.0
aiohttp==3.10.10
python-dateutil==2.9.0
pytz==2024.2
